pyproject.toml
[project]
name = "mailinflow"
version = "0.2.0"
description = "Enhanced ML-based email triage system with multi-account support, spam detection, and auto-unsubscribe"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imap-tools>=1.5.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "openai>=1.0.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]
